+++ src/main/java/com/invadermonky/justenoughmagiculture/configs/mods/JEMConfigSpecialMobs.java
@@ -1,0 +1,23 @@
+package com.invadermonky.justenoughmagiculture.configs.mods;
+
+import com.invadermonky.justenoughmagiculture.JustEnoughMagiculture;
+import com.invadermonky.justenoughmagiculture.util.ModIds.ConstantNames;
+import net.minecraftforge.common.config.Config.Comment;
+import net.minecraftforge.common.config.Config.RequiresMcRestart;
+
+public class JEMConfigSpecialMobs {
+    private static final String MOD_NAME = ConstantNames.SPECIALMOBS;
+    private static final String LANG_KEY = "config." + JustEnoughMagiculture.MOD_ALIAS + ":";
+
+    @RequiresMcRestart
+    @Comment("Enables JEM " + MOD_NAME + " JER mob integration.")
+    public boolean enableJERMobs = true;
+
+    @RequiresMcRestart
+    @Comment("JEM " + MOD_NAME + " Just Enough Resources settings.")
+    public JER JUST_ENOUGH_RESOURCES = new JER();
+    public static class JER {
+        @Comment("Enter the unlocalized name of the mob you wish to remove from JER Integration. Example: DarkCreeper")
+        public String[] disableMobs = new String[0];
+    }
+}